requirements.txt
numpy~=1.24.4
pydantic~=1.10.2
SQLAlchemy~=1.4.41
PyYAML~=6.0
fastapi~=0.47.1
Telethon~=1.38.1
pandas~=2.0.3
ruword_frequency